packages/polaris/polaris_dataset_report/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "polaris_dataset_report"
version = "0.1.0"
description = "Dataset report agent for Polaris"
requires-python = ">=3.11"
dependencies = [
    "polaris",
    "pyyaml",
]

[project.scripts]
dataset-report = "polaris_dataset_report.cli:main"

[project.entry-points."polaris.materializers"]
dataset_report = "polaris_dataset_report.materializers:register_all"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
polaris_dataset_report = ["*.yml"]
